+++ NetworkRequestsProject/app/src/main/java/com/example/networkrequestsproject/domain/UserRepository.kt
@@ -1,0 +1,9 @@
+package com.example.networkrequestsproject.domain
+
+import com.example.networkrequestsproject.domain.model.Person
+import com.example.networkrequestsproject.domain.model.User
+
+interface UserRepository {
+     fun postPerson(person: Person): String?
+     fun getUsers(): List<User>?
+}
