--- NetworkRequestsProject/app/src/main/java/com/example/neworkrequestsproject/domain/UserRepository.kt
@@ -1,9 +1,0 @@
-package com.example.neworkrequestsproject.domain
-
-import com.example.neworkrequestsproject.domain.model.Person
-import com.example.neworkrequestsproject.domain.model.User
-
-interface UserRepository {
-     fun postPerson(person: Person): String?
-     fun getUsers(): List<User>?
-}
