+++ NetworkRequestsProject/app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.NeworkRequestsProject" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.NetworkRequestsProject" parent="Theme.AppCompat.DayNight.DarkActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/toolbar_color</item>
         <item name="colorPrimaryVariant">@color/status_bar_color</item>
